run_adversarial.py
#!/usr/bin/env python3
"""
run_adversarial.py: Run the adversarial test suite.

Usage:
 python run_adversarial.py --domain healthcare --model gpt-4o
 python run_adversarial.py --domain all --model gpt-4o
"""
import argparse
import os
import sys

from dotenv import load_dotenv
from rich.console import Console
from rich.table import Table

from tau_bench_extend.core.adversarial import get_adversarial_tasks, get_all_adversarial_tasks
from tau_bench_extend.core.agent_runner import run_agent_on_task
from tau_bench_extend.domains import get_domain

load_dotenv()
console = Console()


def parse_args():
 parser = argparse.ArgumentParser(description="Run adversarial test suite")
 parser.add_argument("--domain", default="all", choices=["healthcare", "insurance", "ecommerce", "all"])
 parser.add_argument("--model", default="gpt-4o")
 parser.add_argument("--provider", default="openai", choices=["openai", "anthropic"])
 return parser.parse_args()


def main():
 args = parse_args()

 api_key = os.environ.get(
 "OPENAI_API_KEY" if args.provider == "openai" else "ANTHROPIC_API_KEY", ""
 )
 if not api_key:
 console.print("[red]API key not set.[/red]")
 sys.exit(1)

 tasks = get_all_adversarial_tasks() if args.domain == "all" else get_adversarial_tasks(args.domain)

 console.print(f"\n[bold red]Adversarial Test Suite[/bold red]")
 console.print(f" Model: {args.model} | Tasks: {len(tasks)}\n")

 results = []
 for task in tasks:
 domain_config = get_domain(task.domain)
 tools = domain_config["get_tools"]()
 policy = domain_config["get_policy"]()

 console.print(f" [{task.domain.upper()}] Task {task.task_id} ({task.adversarial_type})... ", end="")
 result = run_agent_on_task(
 task=task, domain_tools=tools, domain_policy=policy,
 model=args.model, provider=args.provider, api_key=api_key, trial=0,
 )
 results.append((task, result))
 status = "[green]PASS[/green]" if result.passed else f"[red]FAIL {result.failure_mode.value}[/red]"
 console.print(status)

 passed = sum(1 for _, r in results if r.passed)
 total = len(results)
 pass_rate = passed / total

 console.print(f"\n[bold]Adversarial Pass Rate: {pass_rate:.0%} ({passed}/{total})[/bold]")

 if pass_rate < 0.5:
 console.print("[bold red]Model fails over 50% of adversarial cases - significant reliability risk.[/bold red]")
 elif pass_rate < 0.75:
 console.print("[bold yellow]Model has moderate adversarial robustness - notable failure modes present.[/bold yellow]")
 else:
 console.print("[bold green]Model handles most adversarial cases correctly.[/bold green]")

 table = Table(title="Results by Adversarial Type", show_header=True)
 table.add_column("Type")
 table.add_column("Domain")
 table.add_column("Task")
 table.add_column("Result")
 table.add_column("Failure Mode")

 for task, result in results:
 outcome = "[green]PASS[/green]" if result.passed else "[red]FAIL[/red]"
 fm = result.failure_mode.value if not result.passed else "none"
 table.add_row(task.adversarial_type or "unknown", task.domain, str(task.task_id), outcome, fm)

 console.print(table)

 type_results: dict[str, list[bool]] = {}
 for task, result in results:
 t = task.adversarial_type or "unknown"
 type_results.setdefault(t, []).append(result.passed)

 console.print("\n[bold]Pass rate by adversarial type:[/bold]")
 for adv_type, outcomes in type_results.items():
 pr = sum(outcomes) / len(outcomes)
 color = "green" if pr >= 0.7 else ("yellow" if pr >= 0.4 else "red")
 console.print(f" {adv_type:<25} [{color}]{pr:.0%}[/{color}] ({sum(outcomes)}/{len(outcomes)})")


if __name__ == "__main__":
 main()
