pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tau-bench-extend"
version = "0.1.0"
description = "Extension of Sierra's τ-bench with new domains and consistency scoring"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Jafar", email = "jafar97dev@gmail.com"}
]
keywords = ["llm", "evaluation", "benchmark", "agents", "tau-bench"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "jinja2>=3.0.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[project.scripts]
tau-extend = "tau_bench_extend.cli:main"

[tool.ruff]
line-length = 100
